fix(dsp): keep resampler positions in float64
scale_field() and scale_positions() narrowed the pixel position to
float32 before splitting off its integer part.  A field buffer runs to
about 1e6 samples, where a float32 step is 1/16 of a sample, so every
pixel's interpolation phase was quantised to that grid - and a change of
~1e-5 samples in a line location could flip sparse pixels by a whole
step.  That grid is the "0.016 px" time-base floor recorded against the
interlace-comb work.

The position now stays float64 until coord_int is taken; the fractional
phase and the LUT blend are unchanged.

This CHANGES DECODED OUTPUT.  Measured on
testdata/radius/domesday-ds1-community-north-outer.ldf (--PAL
--exact-speculation), new .tbc against old:

    89.7% of samples differ, mean |delta| 21.2 LSB, max 3829 LSB

The large maxima are the point of the fix: they are the pixels that were
landing a whole float32 position step away from where they belonged.

test_scale_positions_keeps_its_precision_deep_into_the_buffer is the
regression test - the same fractional positions on the same ramp, near
the start of a buffer and 2^20 samples in, must agree; at float32 they
differed by ~0.19 on a slope-3 ramp.  The float32 reference kernel in
test_sinc_lut is updated to keep its position exact too, so it isolates
interpolation error rather than position error.

+++ lddecode/dsp.py
@@ -139,8 +139,12 @@ def scale_field(
         level_adjust = level_adjusts[i]
 
         # reconstructs the waveform at the proper fractional sample position, undoing wow-induced
-        # timing variations
-        coord = np.float32(interpolated_pixel_locs[i])
+        # timing variations.  The position stays float64 until its integer
+        # part is split off: a field buffer runs to ~1e6 samples, where a
+        # float32 step is 1/16 of a sample, and rounding the position there
+        # quantised every pixel's phase to that step (and made a 1e-5-sample
+        # change in a line location flip sparse pixels by a whole step).
+        coord = interpolated_pixel_locs[i]
         coord_int = int(coord)
 
         # fractional phase
@@ -197,7 +201,7 @@ def scale_positions(buf, dsout, pixel_locs, wowfactors, sinc_lut,
     half_taps_m1 = (sinc_tap_count // 2) - 1
 
     for i in range(len(dsout)):
-        coord = np.float32(pixel_locs[i])
+        coord = pixel_locs[i]            # float64: see scale_field
         coord_int = int(coord)
         frac = coord - coord_int
